+++ backend/src/main/java/com/swtp4/backend/repositories/applicationDtos/SubmittedStudentModule.java
@@ -3,10 +3,10 @@ package com.swtp4.backend.repositories.applicationDtos;
 public record SubmittedStudentModule(
         String moduleNumber,
         String moduleName,
-        Integer credits,
+        Long credits,
         String university,
         String moduleMajor,
         String commentStudent,
-        Integer frontend_module_key
+        Long frontend_module_key
 ) {
 }
